pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackbrowserdata"
version = "1.0.0"
description = "Python version of HackBrowserData for extracting browser data"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Ramunezaza"}
]
maintainers = [
    {name = "Ramunezaza"}
]
keywords = ["browser", "security", "forensics", "password", "extraction", "hacking", "cybersecurity"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities"
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0,<9.0.0",
    "cryptography>=43.0.0,<46.0.0", 
    "pycryptodome>=3.20.0,<4.0.0",
    "psutil>=5.8.0,<8.0.0",
    "python-dateutil>=2.8.0,<3.0.0",
    "colorama>=0.4.4,<1.0.0",
    "pywin32>=311; sys_platform == 'win32'",
    "six>=1.16.0,<2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0", 
    "flake8>=6.0.0,<7.0.0",
    "black>=23.0.0,<24.0.0",
    "mypy>=1.0.0,<2.0.0"
]

[project.urls]
Homepage = "https://github.com/ramune0144/HackBrowserDataPython"
Repository = "https://github.com/ramune0144/HackBrowserDataPython"
"Bug Reports" = "https://github.com/ramune0144/HackBrowserDataPython/issues"
"Original Project" = "https://github.com/moonD4rk/HackBrowserData"

[project.scripts]
hackbrowserdata = "hackbrowserdata.hack_browser_data:main"

[tool.setuptools]
packages = ["hackbrowserdata", "hackbrowserdata.browsers", "hackbrowserdata.core", "hackbrowserdata.extractors", "hackbrowserdata.utils"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yml", "*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
